agent: Fix endpoint restore with unmounted BPF filesystem
The function Daemon.restoreOldEndpoints() had errored out when the endpoint BPF
map could not be opened. This would prevent restoring of endpoints when the BPF
filesystem was not mounted. The error does not matter if we can't open the map
as the map will be initialized from scratch and no cleaning up of old
potentially stale entries is necessary.

Fixes: 4ecf111c35f ("agent: Fix temporary corruption of BPF endpoint map on restart")

+++ daemon/state.go
@@ -72,7 +72,7 @@ func (d *Daemon) restoreOldEndpoints(dir string, clean bool) (*endpointRestoreSt
 
 	existingEndpoints, err := lxcmap.DumpToMap()
 	if err != nil {
-		return state, err
+		log.WithError(err).Warning("Unable to open endpoint map while restoring. Skipping cleanup of endpoint map on startup")
 	}
 
 	dirFiles, err := ioutil.ReadDir(dir)
@@ -144,8 +144,10 @@ func (d *Daemon) restoreOldEndpoints(dir string, clean bool) (*endpointRestoreSt
 
 		state.restored = append(state.restored, ep)
 
-		delete(existingEndpoints, ep.IPv4.String())
-		delete(existingEndpoints, ep.IPv6.String())
+		if existingEndpoints != nil {
+			delete(existingEndpoints, ep.IPv4.String())
+			delete(existingEndpoints, ep.IPv6.String())
+		}
 	}
 
 	log.WithFields(logrus.Fields{
@@ -153,12 +155,14 @@ func (d *Daemon) restoreOldEndpoints(dir string, clean bool) (*endpointRestoreSt
 		"count.total":    len(possibleEPs),
 	}).Info("Endpoints restored")
 
-	for hostIP, info := range existingEndpoints {
-		if ip := net.ParseIP(hostIP); !info.IsHost() && ip != nil {
-			if err := lxcmap.DeleteEntry(ip); err != nil {
-				log.WithError(err).Warn("Unable to delete obsolete endpoint from BPF map")
-			} else {
-				log.Debugf("Removed outdated endpoint %d from endpoint map", info.LxcID)
+	if existingEndpoints != nil {
+		for hostIP, info := range existingEndpoints {
+			if ip := net.ParseIP(hostIP); !info.IsHost() && ip != nil {
+				if err := lxcmap.DeleteEntry(ip); err != nil {
+					log.WithError(err).Warn("Unable to delete obsolete endpoint from BPF map")
+				} else {
+					log.Debugf("Removed outdated endpoint %d from endpoint map", info.LxcID)
+				}
 			}
 		}
 	}
